Alterations to RDM values and models affected by name changes
To facilitate the name changes expected to the Heritage Asset and Heritage Area graphs 4 concepts have changed in the RDM and been updated in the relevant branches in the 2 relevant graphs.

The vocabularies affected are Name_Type_20200515.xml and Resource_Model_Types_20200521.xml.

The concepts in Name Type affected are Heritage Asset Name and Heritage Area Name. These are now Monument Name and Area Name.

They are used as default values in the Heritage Asset Name Type and Heritage Area Name Type nodes on the Names branch in their respective models. Both branches have been checked and updated.

The other concepts are Heritage Asset and Heritage Area in the Resource_Models vocab. These are now Monument and Area respectively and the Resource Model Type branch has been checked and updated in both models.

This push includes the updated vocabularies for Name Type and Resource Type as well as the latest version of Collections.xml